requirements.txt
ffpyplayer
numpy
pillow